moved my thankyou function to its own component
