+++ code/src/App.js
@@ -1,10 +1,11 @@
 import React, { useState } from 'react';
-import { Name } from 'components/Name';
-import { Hear } from 'components/Hear';
-import { Overall } from 'components/Overall';
-import { Delivery } from 'components/Delivery';
-import { Purchase } from 'components/Purchase';
-import { Rate } from 'components/Rate';
+import { Name } from 'components/Name/Name';
+import { Hear } from 'components/Hear/Hear';
+import { Overall } from 'components/Overall/Overall';
+import { Delivery } from 'components/Delivery/Delivery';
+import { Purchase } from 'components/Purchase/Purchase';
+import { Rate } from 'components/Rate/Rate';
+import { ThankYou } from 'components/Thanks';
 
 export const App = () => {
   const [step, setStep] = useState(1);
@@ -30,22 +31,6 @@ export const App = () => {
     // Set the step back to the first question
     setStep(1);
   };
-
-  const thankYouMessage = (
-    <div>
-      <h2>Thank you for participating in the survey!</h2>
-      <p>Here is a summary of your responses:</p>
-      <ul>
-        <li>Name: {name}</li>
-        <li>How did you hear about us?: {hear}</li>
-        <li>Overall satisfaction: {overall}</li>
-        <li>Delivery satisfaction: {delivery}</li>
-        <li>Purchase satisfaction: {purchase}</li>
-        <li>How likely are you to recommend us?: {rate}</li>
-      </ul>
-      <button type="button" onClick={handleReset}>Take the survey again</button>
-    </div>
-  );
 
   return (
     <>
@@ -85,12 +70,19 @@ export const App = () => {
           setRate={setRate}
           step={step} />
       )}
-
       {step < 7 && (
-        <button type="button" onClick={handleStepIncrease}>Next step</button>
+        <button type="button" onClick={handleStepIncrease}>Next question</button>
       )}
-
-      {step === 7 && thankYouMessage}
+      {step === 7 && (
+        <ThankYou
+          name={name}
+          hear={hear}
+          overall={overall}
+          delivery={delivery}
+          purchase={purchase}
+          rate={rate}
+          handleReset={handleReset} />
+      )}
     </>
   );
 };
