+++ code/src/components/Rate/Rate.js
